+++ src/RepetierServerSharpApi/Models/Config/RepetierPrinterConfigMovement.cs
@@ -8,31 +8,31 @@ namespace AndreasReitberger.API.Repetier.Models
         #region Properties
         [ObservableProperty]
         [JsonProperty("G10Distance"), JsonPropertyName("G10Distance")]
-        public partial long G10Distance { get; set; }
+        public partial double G10Distance { get; set; }
 
         [ObservableProperty]
         [JsonProperty("G10LongDistance"), JsonPropertyName("G10LongDistance")]
-        public partial long G10LongDistance { get; set; }
+        public partial double G10LongDistance { get; set; }
 
         [ObservableProperty]
         [JsonProperty("G10Speed"), JsonPropertyName("G10Speed")]
-        public partial long G10Speed { get; set; }
+        public partial double G10Speed { get; set; }
 
         [ObservableProperty]
         [JsonProperty("G10ZLift"), JsonPropertyName("G10ZLift")]
-        public partial long G10ZLift { get; set; }
+        public partial double G10ZLift { get; set; }
 
         [ObservableProperty]
         [JsonProperty("G11ExtraDistance"), JsonPropertyName("G11ExtraDistance")]
-        public partial long G11ExtraDistance { get; set; }
+        public partial double G11ExtraDistance { get; set; }
 
         [ObservableProperty]
         [JsonProperty("G11ExtraLongDistance"), JsonPropertyName("G11ExtraLongDistance")]
-        public partial long G11ExtraLongDistance { get; set; }
+        public partial double G11ExtraLongDistance { get; set; }
 
         [ObservableProperty]
         [JsonProperty("G11Speed"), JsonPropertyName("G11Speed")]
-        public partial long G11Speed { get; set; }
+        public partial double G11Speed { get; set; }
 
         [ObservableProperty]
         [JsonProperty("allEndstops"), JsonPropertyName("allEndstops")]
@@ -44,15 +44,15 @@ namespace AndreasReitberger.API.Repetier.Models
 
         [ObservableProperty]
         [JsonProperty("defaultAcceleration"), JsonPropertyName("defaultAcceleration")]
-        public partial long DefaultAcceleration { get; set; }
+        public partial double DefaultAcceleration { get; set; }
 
         [ObservableProperty]
         [JsonProperty("defaultRetractAcceleration"), JsonPropertyName("defaultRetractAcceleration")]
-        public partial long DefaultRetractAcceleration { get; set; }
+        public partial double DefaultRetractAcceleration { get; set; }
 
         [ObservableProperty]
         [JsonProperty("defaultTravelAcceleration"), JsonPropertyName("defaultTravelAcceleration")]
-        public partial long DefaultTravelAcceleration { get; set; }
+        public partial double DefaultTravelAcceleration { get; set; }
 
         [ObservableProperty]
         [JsonProperty("invertX"), JsonPropertyName("invertX")]
@@ -68,11 +68,11 @@ namespace AndreasReitberger.API.Repetier.Models
 
         [ObservableProperty]
         [JsonProperty("maxXYSpeed"), JsonPropertyName("maxXYSpeed")]
-        public partial long MaxXySpeed { get; set; }
+        public partial double MaxXySpeed { get; set; }
 
         [ObservableProperty]
         [JsonProperty("maxZSpeed"), JsonPropertyName("maxZSpeed")]
-        public partial long MaxZSpeed { get; set; }
+        public partial double MaxZSpeed { get; set; }
 
         [ObservableProperty]
         [JsonProperty("movebuffer"), JsonPropertyName("movebuffer")]
@@ -92,31 +92,31 @@ namespace AndreasReitberger.API.Repetier.Models
 
         [ObservableProperty]
         [JsonProperty("xHome"), JsonPropertyName("xHome")]
-        public partial long XHome { get; set; }
+        public partial double XHome { get; set; }
 
         [ObservableProperty]
         [JsonProperty("xMax"), JsonPropertyName("xMax")]
-        public partial long XMax { get; set; }
+        public partial double XMax { get; set; }
 
         [ObservableProperty]
         [JsonProperty("xMin"), JsonPropertyName("xMin")]
-        public partial long XMin { get; set; }
+        public partial double XMin { get; set; }
 
         [ObservableProperty]
         [JsonProperty("xyJerk"), JsonPropertyName("xyJerk")]
-        public partial long XyJerk { get; set; }
+        public partial double XyJerk { get; set; }
 
         [ObservableProperty]
         [JsonProperty("xyPrintAcceleration"), JsonPropertyName("xyPrintAcceleration")]
-        public partial long XyPrintAcceleration { get; set; }
+        public partial double XyPrintAcceleration { get; set; }
 
         [ObservableProperty]
         [JsonProperty("xySpeed"), JsonPropertyName("xySpeed")]
-        public partial long XySpeed { get; set; }
+        public partial double XySpeed { get; set; }
 
         [ObservableProperty]
         [JsonProperty("xyTravelAcceleration"), JsonPropertyName("xyTravelAcceleration")]
-        public partial long XyTravelAcceleration { get; set; }
+        public partial double XyTravelAcceleration { get; set; }
 
         [ObservableProperty]
         [JsonProperty("yEndstop"), JsonPropertyName("yEndstop")]
@@ -124,15 +124,15 @@ namespace AndreasReitberger.API.Repetier.Models
 
         [ObservableProperty]
         [JsonProperty("yHome"), JsonPropertyName("yHome")]
-        public partial long YHome { get; set; }
+        public partial double YHome { get; set; }
 
         [ObservableProperty]
         [JsonProperty("yMax"), JsonPropertyName("yMax")]
-        public partial long YMax { get; set; }
+        public partial double YMax { get; set; }
 
         [ObservableProperty]
         [JsonProperty("yMin"), JsonPropertyName("yMin")]
-        public partial long YMin { get; set; }
+        public partial double YMin { get; set; }
 
         [ObservableProperty]
         [JsonProperty("zEndstop"), JsonPropertyName("zEndstop")]
@@ -140,7 +140,7 @@ namespace AndreasReitberger.API.Repetier.Models
 
         [ObservableProperty]
         [JsonProperty("zHome"), JsonPropertyName("zHome")]
-        public partial long ZHome { get; set; }
+        public partial double ZHome { get; set; }
 
         [ObservableProperty]
         [JsonProperty("zJerk"), JsonPropertyName("zJerk")]
@@ -148,23 +148,23 @@ namespace AndreasReitberger.API.Repetier.Models
 
         [ObservableProperty]
         [JsonProperty("zMax"), JsonPropertyName("zMax")]
-        public partial long ZMax { get; set; }
+        public partial double ZMax { get; set; }
 
         [ObservableProperty]
         [JsonProperty("zMin"), JsonPropertyName("zMin")]
-        public partial long ZMin { get; set; }
+        public partial double ZMin { get; set; }
 
         [ObservableProperty]
         [JsonProperty("zPrintAcceleration"), JsonPropertyName("zPrintAcceleration")]
-        public partial long ZPrintAcceleration { get; set; }
+        public partial double ZPrintAcceleration { get; set; }
 
         [ObservableProperty]
         [JsonProperty("zSpeed"), JsonPropertyName("zSpeed")]
-        public partial long ZSpeed { get; set; }
+        public partial double ZSpeed { get; set; }
 
         [ObservableProperty]
         [JsonProperty("zTravelAcceleration"), JsonPropertyName("zTravelAcceleration")]
-        public partial long ZTravelAcceleration { get; set; }
+        public partial double ZTravelAcceleration { get; set; }
         #endregion
 
         #region Overrides
